% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-extras.R
\name{pull.list}
\alias{pull.list}
\title{Extract a single column from a list of lists}
\usage{
\method{pull}{list}(x, col, ...)
}
\description{
List version of dplyr's pull verb. Note that because there is no type
enforcement for lists of lists, we'll return a vector only if the actual
type of all elements in the column is the same. Otherwise we return a list.
}
